refactor(cli): split workload.go into per-command files
workload.go held 25 commands across ~175 declarations in a single 5305-line
file. Split each command into its own file (apply, create, debug, delete,
describe, edit, exec, explain, export, expose, forward, get, images, import,
install, logs, push, reconcile, rollout, scale, scan, validate, wait, watch),
with cross-cutting helpers (shared command-context types, kubectl-passthrough
wrappers, push/reconcile shared helpers) collected in shared.go and the watch
debounce helpers grouped with watch. workload.go remains the root command.

This is a behavior-preserving move performed by slicing the original source at
top-level declaration boundaries (exact text and doc comments preserved);
per-file imports were rebuilt with goimports. A few interleaved declarations
(errNotDirectory, watchCmdLong, ErrInvalidComplianceThreshold) were relocated
to the command file that owns them.

The relocation surfaced the repeated annotations.AnnotationPermission value
"write" under CI's new-issue goconst gate; extracted it into a permissionWrite
constant.
