+++ pkg/cli/cmd/workload/apply.go
@@ -1,0 +1,23 @@
+package workload
+
+import (
+	"github.com/devantler-tech/ksail/v7/pkg/cli/annotations"
+	"github.com/devantler-tech/ksail/v7/pkg/client/kubectl"
+	"github.com/spf13/cobra"
+)
+
+// NewApplyCmd creates the workload apply command.
+func NewApplyCmd() *cobra.Command {
+	cmd := newKubectlCommand(func(client *kubectl.Client, kubeconfigPath string) *cobra.Command {
+		return client.CreateApplyCommand(kubeconfigPath)
+	})
+
+	// Mark as requiring permission for edit operations.
+	// Note: kubectl/helm may have their own confirmation prompts for certain operations.
+	// The permission system here is for AI tool execution confirmation.
+	cmd.Annotations = map[string]string{
+		annotations.AnnotationPermission: permissionWrite,
+	}
+
+	return cmd
+}
