+++ pkg/cli/cmd/workload/describe.go
@@ -1,0 +1,13 @@
+package workload
+
+import (
+	"github.com/devantler-tech/ksail/v7/pkg/client/kubectl"
+	"github.com/spf13/cobra"
+)
+
+// NewDescribeCmd creates the workload describe command.
+func NewDescribeCmd() *cobra.Command {
+	return newKubectlCommand(func(client *kubectl.Client, kubeconfigPath string) *cobra.Command {
+		return client.CreateDescribeCommand(kubeconfigPath)
+	})
+}
