+++ pkg/cli/cmd/workload/explain.go
@@ -1,0 +1,13 @@
+package workload
+
+import (
+	"github.com/devantler-tech/ksail/v7/pkg/client/kubectl"
+	"github.com/spf13/cobra"
+)
+
+// NewExplainCmd creates the workload explain command.
+func NewExplainCmd() *cobra.Command {
+	return newKubectlCommand(func(client *kubectl.Client, kubeconfigPath string) *cobra.Command {
+		return client.CreateExplainCommand(kubeconfigPath)
+	})
+}
