+++ pkg/cli/cmd/workload/get.go
@@ -1,0 +1,13 @@
+package workload
+
+import (
+	"github.com/devantler-tech/ksail/v7/pkg/client/kubectl"
+	"github.com/spf13/cobra"
+)
+
+// NewGetCmd creates the workload get command.
+func NewGetCmd() *cobra.Command {
+	return newKubectlCommand(func(client *kubectl.Client, kubeconfigPath string) *cobra.Command {
+		return client.CreateGetCommand(kubeconfigPath)
+	})
+}
