+++ pkg/cli/cmd/workload/logs.go
@@ -1,0 +1,13 @@
+package workload
+
+import (
+	"github.com/devantler-tech/ksail/v7/pkg/client/kubectl"
+	"github.com/spf13/cobra"
+)
+
+// NewLogsCmd creates the workload logs command.
+func NewLogsCmd() *cobra.Command {
+	return newKubectlCommand(func(client *kubectl.Client, kubeconfigPath string) *cobra.Command {
+		return client.CreateLogsCommand(kubeconfigPath)
+	})
+}
