+++ pkg/cli/cmd/workload/scale.go
@@ -1,0 +1,20 @@
+package workload
+
+import (
+	"github.com/devantler-tech/ksail/v7/pkg/cli/annotations"
+	"github.com/devantler-tech/ksail/v7/pkg/client/kubectl"
+	"github.com/spf13/cobra"
+)
+
+// NewScaleCmd creates the workload scale command.
+func NewScaleCmd() *cobra.Command {
+	cmd := newKubectlCommand(func(client *kubectl.Client, kubeconfigPath string) *cobra.Command {
+		return client.CreateScaleCommand(kubeconfigPath)
+	})
+
+	cmd.Annotations = map[string]string{
+		annotations.AnnotationPermission: permissionWrite,
+	}
+
+	return cmd
+}
